修正1

+++ app/controllers/items_controller.rb
@@ -34,11 +34,13 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+   if  current_user.id == @item.user_id
     if @item.destroy
       redirect_to root_path
     else
-      render :new
+      render :show
     end
+   end
   end
 
   def set_item
